+++ packages/babel-plugin-transform-react-jsx/package.json
@@ -16,10 +16,11 @@
     "babel-plugin"
   ],
   "dependencies": {
-    "@babel/helper-builder-react-jsx": "workspace:^7.10.4",
-    "@babel/helper-builder-react-jsx-experimental": "workspace:^7.12.11",
+    "@babel/helper-annotate-as-pure": "workspace:^7.12.10",
+    "@babel/helper-module-imports": "workspace:^7.12.5",
     "@babel/helper-plugin-utils": "workspace:^7.10.4",
-    "@babel/plugin-syntax-jsx": "workspace:^7.12.1"
+    "@babel/plugin-syntax-jsx": "workspace:^7.12.1",
+    "@babel/types": "workspace:^7.12.11"
   },
   "peerDependencies": {
     "@babel/core": "^7.0.0-0"
