+++ packages/babel-plugin-transform-react-jsx/src/development.js
@@ -1,0 +1,6 @@
+import createPlugin from "./create-plugin.js";
+
+export default createPlugin({
+  name: "transform-react-jsx/development",
+  development: true,
+});
